@@ -1,12 +1,12 @@
 [tool.poetry]
 name = "manobase"
-version = "0.9.0"
+version = "1.0.0"
 description = "Pishahang MANO framework base"
 authors = ["Manuel Peuster <manuel.peuster@upb.de>"]
 license = "Apache-2.0"
 
 [tool.poetry.dependencies]
-python = "^3.6"
+python = "^3.7"
 amqpstorm = "^2.7.2"
 PyYAML = "^5.3.1"
 requests = "^2.23.0"
