@@ -6,7 +6,7 @@ authors = ["Hadi Razzaghi Kouchaksaraei <hadi.razzaghi@upb.de>"]
 license = "apache-2.0"
 
 [tool.poetry.dependencies]
-python = "^3.6"
+python = "^3.7"
 pycrypto = "2.6.1"
 manobase = { path = "../../base" }
 pyzmq = "^19.0.0"
