@@ -1,12 +1,12 @@
 [tool.poetry]
 name = "slm"
-version = "0.9.0"
+version = "0.9.1"
 description = "Pishahang Service Lifecycle Management Plugin"
 authors = ["Pishahang <pishahang@noreply.github.com>"]
 license = "apache-2.0"
 
 [tool.poetry.dependencies]
-python = "^3.6"
+python = "^3.7"
 pycrypto = "2.6.1"
 manobase = { path = "../../base" }
 
