@@ -6,8 +6,7 @@ authors = ["Pishahang <pishahang@noreply.github.com>"]
 license = "apache-2.0"
 
 [tool.poetry.dependencies]
-python = "^3.6"
-amqpstorm = "^2.7.2"
+python = "^3.7"
 manobase = { path = "../../base" }
 
 [tool.poetry.dev-dependencies]
