@@ -1,5 +1,5 @@
 """
-Copyright (c) 2015 SONATA-NFV
+Copyright (c) 2015 SONATA-NFV, 2017 Pishahang
 ALL RIGHTS RESERVED.
 
 Licensed under the Apache License, Version 2.0 (the "License");
@@ -14,12 +14,12 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 
-Neither the name of the SONATA-NFV [, ANY ADDITIONAL AFFILIATION]
+Neither the name of the SONATA-NFV, Pishahang,
 nor the names of its contributors may be used to endorse or promote
 products derived from this software without specific prior written
 permission.
 
-This work has been performed in the framework of the SONATA project,
+Parts of this work have been performed in the framework of the SONATA project,
 funded by the European Commission under Grant number 671517 through
 the Horizon 2020 and 5G-PPP programmes. The authors would like to
 acknowledge the contributions of their colleagues of the SONATA
@@ -27,19 +27,20 @@ partner consortium (www.sonata-nfv.eu).
 """
 
 import logging
-import json
-import time
 import os
-import threading
+import time
+from threading import Event, Thread
+from uuid import uuid4
 
-from manobase import messaging
+from amqpstorm import AMQPConnectionError
+
+from manobase.messaging import AsyncioBrokerConnection, Message
 
 logging.basicConfig(level=logging.INFO)
 LOG = logging.getLogger("manobase:plugin")
-LOG.setLevel(logging.DEBUG)
 
 
-class ManoBasePlugin(object):
+class ManoBasePlugin:
     """
     Abstract class that should be inherited by other MANO plugins.
     This class provides basic mechanisms to
@@ -48,18 +49,22 @@ class ManoBasePlugin(object):
     - send/receive notifications
     - register / de-register plugin to plugin manager
 
-    It also implements a automatic heartbeat mechanism that periodically sends
+    It also implements an automatic heartbeat mechanism that periodically sends
     heartbeat notifications.
     """
 
-    def __init__(self,
-                 name="son-plugin",
-                 version=None,
-                 description=None,
-                 auto_register=True,
-                 wait_for_registration=True,
-                 start_running=True,
-                 auto_heartbeat_rate=0.5):
+    def __init__(
+        self,
+        name="son-plugin",
+        version=None,
+        description=None,
+        auto_register=True,
+        wait_for_registration=True,
+        start_running=True,
+        auto_heartbeat_rate=0.5,
+        use_loopback_connection=False,
+        fake_registration=False,
+    ):
         """
         Performs plugin initialization steps, e.g., connection setup
         :param name: Plugin name prefix
@@ -68,36 +73,44 @@ class ManoBasePlugin(object):
         :param auto_register: Automatically register on init
         :param wait_for_registration: Wait for registration before returning from init
         :param auto_heartbeat_rate: rate of automatic heartbeat notifications 1/n seconds. 0=deactivated
-        :return:
+        :param use_loopback_connection: For unit testing: Whether to set up the connection with `is_loopback=True`
+        :param fake_registration: For unit testing: Whether to fake the registration process without a real PluginManager instance involved
         """
         self.name = "%s.%s" % (name, self.__class__.__name__)
         self.version = version
         self.description = description
-        self.uuid = None  # uuid given by plugin manager on registration
-        self.state = None  # the state of this plugin READY/RUNNING/PAUSED/FAILED
+        self.uuid: str = None  # uuid given by plugin manager on registration
+        self.state: str = None  # the state of this plugin READY/RUNNING/PAUSED/FAILED
+        self._fake_registration = fake_registration
 
-        LOG.info(
-            "Starting MANO Plugin: %r ..." % self.name)
-        # create and initialize broker connection
+        self._registered_event = Event()
+
+        LOG.info("Starting MANO Plugin: %s ...", self.name)
+
+        # Initialize broker connection
         while True:
             try:
-                self.manoconn = messaging.ManoBrokerRequestResponseConnection(self.name)
+                self.conn = AsyncioBrokerConnection(
+                    self.name, is_loopback=use_loopback_connection
+                )
                 break
-            except:
+            except AMQPConnectionError:
                 time.sleep(5)
-        # register subscriptions
+        self.manoconn = self.conn  # For backwards compatibility
+
         LOG.info("Plugin connected to broker.")
 
         self.declare_subscriptions()
-        # register to plugin manager
+
+        # Register at plugin manager
         if auto_register:
-            while self.uuid is None:
-                self.register()
-                if wait_for_registration:
-                    self._wait_for_registration()
-        # kick-off automatic heartbeat mechanism
-        self._auto_heartbeat(auto_heartbeat_rate)
-        # jump to run
+            self.register()
+            if wait_for_registration:
+                self._wait_for_registration()
+
+        # Start hearbeat mechanism
+        self._start_heartbeats(auto_heartbeat_rate)
+
         if start_running:
             self.run()
 
@@ -108,11 +121,9 @@ class ManoBasePlugin(object):
         """
         # de-register this plugin
         self.deregister()
-        self.manoconn.stop_connection()
-        self.manoconn.stop_threads()
-        del self.manoconn
+        self.conn.close()
 
-    def _auto_heartbeat(self, rate):
+    def _start_heartbeats(self, rate):
         """
         A simple periodic heartbeat mechanism.
         (much room for improvements here)
@@ -126,18 +137,15 @@ class ManoBasePlugin(object):
             while True:
                 if self.uuid is not None:
                     self._send_heartbeat()
-                time.sleep(1/rate)
+                time.sleep(1 / rate)
 
-        # run heartbeats in separated thread
-        t = threading.Thread(target=run)
-        t.daemon = True
-        t.start()
+        Thread(target=run, daemon=True).start()
 
     def _send_heartbeat(self):
-        self.manoconn.notify(
+        self.conn.notify(
             "platform.management.plugin.%s.heartbeat" % str(self.uuid),
-            json.dumps({"uuid": self.uuid,
-            "state": str(self.state)}))
+            {"uuid": self.uuid, "state": str(self.state)},
+        )
 
     def declare_subscriptions(self):
         """
@@ -145,9 +153,10 @@ class ManoBasePlugin(object):
         But: The this superclass method should be called in any case.
         """
         # plugin status update subscription
-        self.manoconn.register_notification_endpoint(
+        self.conn.register_notification_endpoint(
             self.on_plugin_status_update,  # call back method
-            "platform.management.plugin.status")
+            "platform.management.plugin.status",
+        )
 
     def run(self):
         """
@@ -157,21 +166,21 @@ class ManoBasePlugin(object):
         while True:
             time.sleep(1)
 
-    def on_lifecycle_start(self, ch, method, properties, message):
+    def on_lifecycle_start(self, message: Message):
         """
         To be overwritten by subclass
         """
         LOG.debug("Received lifecycle.start event.")
         self.state = "RUNNING"
 
-    def on_lifecycle_pause(self, ch, method, properties, message):
+    def on_lifecycle_pause(self, message: Message):
         """
         To be overwritten by subclass
         """
         LOG.debug("Received lifecycle.pause event.")
         self.state = "PAUSED"
 
-    def on_lifecycle_stop(self, ch, method, properties, message):
+    def on_lifecycle_stop(self, message: Message):
         """
         To be overwritten by subclass
         """
@@ -186,97 +195,92 @@ class ManoBasePlugin(object):
         LOG.debug("Received registration ok event.")
         pass
 
-    def on_plugin_status_update(self, ch, method, properties, message):
+    def on_plugin_status_update(self, message: Message):
         """
         To be overwritten by subclass.
         Called when a plugin list status update
         is received from the plugin manager.
         """
-        LOG.debug("Received plugin status update %r." % str(message))
+        LOG.debug("Received plugin status update %s.", message.payload)
 
     def register(self):
         """
-        Send a register request to the plugin manager component to announce this plugin.
+        Register this plugin at the plugin manager.
         """
-        message = {"name": self.name,
-                   "version": self.version,
-                   "description": self.description}
+        self.conn.run_coroutine(self._register())
 
-        self.manoconn.call_async(self._on_register_response,
-                                 "platform.management.plugin.register",
-                                 json.dumps(message))
+    async def _register(self):
+        if self._fake_registration:
+            self.uuid = str(uuid4())
+        else:
+            response = (
+                await self.conn.call(
+                    "platform.management.plugin.register",
+                    {
+                        "name": self.name,
+                        "version": self.version,
+                        "description": self.description,
+                    },
+                )
+            ).payload
 
-    def _on_register_response(self, ch, method, props, response):
-        """
-        Event triggered when register response is received.
-        :param props: response properties
-        :param response: response body
-        :return: None
-        """
-        response = json.loads(str(response))
-        if response.get("status") != "OK":
-            LOG.debug("Response %r" % response)
-            LOG.error("Plugin registration failed. Exit.")
-            exit(1)
-        self.uuid = response.get("uuid")
-        # mark this plugin to be ready to be started
+            if response["status"] != "OK":
+                LOG.debug("Response %r", response)
+                LOG.error("Plugin registration failed. Exit.")
+                exit(1)
+
+            self.uuid = response["uuid"]
+
         self.state = "READY"
-        LOG.info("Plugin registered with UUID: %r" % response.get("uuid"))
-        # jump to on_registration_ok()
+
+        LOG.info("Plugin registered with UUID: %s", self.uuid)
         self.on_registration_ok()
-        # subscribe to start topic
+
         self._register_lifecycle_endpoints()
-        # start heartbeat mechanism
-        self._send_heartbeat()
+        self._registered_event.set()
 
     def deregister(self):
         """
-        Send a deregister event to the plugin manager component.
+        De-registers the plugin at the plugin manager.
         """
+        self.conn.run_coroutine(self._deregister())
+
+    async def _deregister(self):
         LOG.info("De-registering plugin...")
-        message = {"uuid": self.uuid}
-        self.manoconn.call_async(self._on_deregister_response,
-                                 "platform.management.plugin.deregister",
-                                 json.dumps(message))
-    def _on_deregister_response(self, ch, method, props, response):
-        """
-        Event triggered when de-register response is received.
-        :param props: response properties
-        :param response: response body
-        :return: None
-        """
-        response = json.loads(str(response))
-        if response.get("status") != "OK":
+
+        response = await self.conn.call(
+            "platform.management.plugin.deregister", {"uuid": self.uuid},
+        )
+
+        if response.payload["status"] != "OK":
             LOG.error("Plugin de-registration failed. Exit.")
             exit(1)
+        self.uuid = None
+        self._registered_event.clear()
         LOG.info("Plugin de-registered.")
 
-    def _wait_for_registration(self, timeout=5, sleep_interval=0.1):
+    def _wait_for_registration(self, timeout=5):
         """
-        Method to do active waiting until the registration is completed.
-        (not nice, but ok for now)
-        :param timeout: max wait
-        :param sleep_interval: sleep interval
+        Method to block until the registration is completed or a timeout reached.
+        :param timeout: Timeout in seconds
         :return: None
         """
-        # FIXME: Use threading.Event() for this?
-        c = 0
-        LOG.debug("Waiting for registration (timeout=%d) ..." % timeout)
-        while self.uuid is None and c < timeout:
-            time.sleep(sleep_interval)
-            c += sleep_interval
+        LOG.debug("Waiting for registration (timeout=%d) ...", timeout)
+        self._registered_event.wait(timeout)
 
     def _register_lifecycle_endpoints(self):
         if self.uuid is not None:
+            base_topic = "platform.management.plugin.%s.lifecycle." % str(self.uuid)
+
             # lifecycle.start
-            self.manoconn.register_notification_endpoint(
-                self.on_lifecycle_start,  # call back method
-                "platform.management.plugin.%s.lifecycle.start" % str(self.uuid))
+            self.conn.register_notification_endpoint(
+                self.on_lifecycle_start, base_topic + "start",
+            )
             # lifecycle.pause
-            self.manoconn.register_notification_endpoint(
-                self.on_lifecycle_pause,  # call back method
-                "platform.management.plugin.%s.lifecycle.pause" % str(self.uuid))
+            self.conn.register_notification_endpoint(
+                self.on_lifecycle_pause, base_topic + "pause",
+            )
             # lifecycle.stop
-            self.manoconn.register_notification_endpoint(
-                self.on_lifecycle_stop,  # call back method
-                "platform.management.plugin.%s.lifecycle.stop" % str(self.uuid))
+            self.conn.register_notification_endpoint(
+                self.on_lifecycle_stop, base_topic + "stop",
+            )
