@@ -24,14 +24,12 @@ funded by the European Commission under Grant number 671517 through
 the Horizon 2020 and 5G-PPP programmes. The authors would like to
 acknowledge the contributions of their colleagues of the SONATA
 partner consortium (www.sonata-nfv.eu).
+
+Adds a REST interface to the plugin manager to control plugins registered to the platform.
 """
 
-"""
-Adds a REST interface to the plugin manager to control plugins registered to the platfoem.
-"""
 import logging
 import threading
-import json
 from flask import Flask, request
 import flask_restful as fr
 from mongoengine import DoesNotExist
@@ -44,25 +42,23 @@ logging.getLogger("werkzeug").setLevel(logging.WARNING)
 
 
 class PluginsEndpoint(fr.Resource):
-
     def get(self):
         LOG.debug("GET plugin list")
         return [p.uuid for p in model.Plugin.objects], 200
 
 
 class PluginEndpoint(fr.Resource):
-
     def get(self, plugin_uuid=None):
-        LOG.debug("GET plugin info for: %r" % plugin_uuid)
+        LOG.debug("GET plugin info for: %r", plugin_uuid)
         try:
             p = model.Plugin.objects.get(uuid=plugin_uuid)
             return p.to_dict(), 200
         except DoesNotExist:
-            LOG.error("Lookup error: %r" % plugin_uuid)
+            LOG.error("Lookup error: %r", plugin_uuid)
             return {}, 404
 
     def delete(self, plugin_uuid=None):
-        LOG.debug("DELETE plugin: %r" % plugin_uuid)
+        LOG.debug("DELETE plugin: %r", plugin_uuid)
         try:
             p = model.Plugin.objects.get(uuid=plugin_uuid)
             # send lifecycle stop event to plugin
@@ -70,20 +66,19 @@ class PluginEndpoint(fr.Resource):
             # TODO ensure that record is deleted even if plugin does not deregister itself (use a timeout?)
             return {}, 200
         except DoesNotExist:
-            LOG.error("Lookup error: %r" % plugin_uuid)
+            LOG.error("Lookup error: %r", plugin_uuid)
             return {}, 404
 
 
 class PluginLifecycleEndpoint(fr.Resource):
-
     def put(self, plugin_uuid=None):
-        LOG.debug("PUT plugin lifecycle: %r" % plugin_uuid)
+        LOG.debug("PUT plugin lifecycle: %r", plugin_uuid)
         try:
             p = model.Plugin.objects.get(uuid=plugin_uuid)
             # get target state from request body
             ts = request.get_json().get("target_state", None)
             if ts is None:
-                LOG.error("Malformed request: %r" % request.json)
+                LOG.error("Malformed request: %r", request.json)
                 return {"message": "malformed request"}, 500
             if ts == "start":
                 PM.send_start_notification(p)
@@ -95,7 +90,7 @@ class PluginLifecycleEndpoint(fr.Resource):
                 return {"message": "Malformed request"}, 500
             return {}, 200
         except DoesNotExist:
-            LOG.error("Lookup error: %r" % plugin_uuid)
+            LOG.error("Lookup error: %r", plugin_uuid)
             return {}, 404
 
 
@@ -112,11 +107,12 @@ api.add_resource(PluginLifecycleEndpoint, "/api/plugins/<string:plugin_uuid>/lif
 
 def _start_flask(host, port):
     # start the Flask server (not the best performance but ok for our use case)
-    app.run(host=host,
-            port=port,
-            debug=True,
-            use_reloader=False  # this is needed to run Flask in a non-main thread
-            )
+    app.run(
+        host=host,
+        port=port,
+        debug=True,
+        use_reloader=False,  # this is needed to run Flask in a non-main thread
+    )
 
 
 def start(pm, host="0.0.0.0", port=8001):
@@ -125,4 +121,4 @@ def start(pm, host="0.0.0.0", port=8001):
     thread = threading.Thread(target=_start_flask, args=(host, port))
     thread.daemon = True
     thread.start()
-    LOG.info("Started management REST interface @ http://%s:%d" % (host, port))
+    LOG.info("Started management REST interface @ http://%s:%d", host, port)
