@@ -27,5 +27,6 @@ from flm import flm
 def main():
     flm.main()
 
-if __name__ == '__main__':
+
+if __name__ == "__main__":
     main()
