@@ -27,5 +27,6 @@ from sdn import sdn
 def main():
     sdn.main()
 
-if __name__ == '__main__':
+
+if __name__ == "__main__":
     main()
