@@ -27,5 +27,6 @@ from slm import slm
 def main():
     slm.main()
 
-if __name__ == '__main__':
+
+if __name__ == "__main__":
     main()
