@@ -4,5 +4,6 @@ from specific_manager_registry import specificmanagerregistry
 def main():
     specificmanagerregistry.main()
 
-if __name__ == '__main__':
+
+if __name__ == "__main__":
     main()
